ADD: Cleaned up group pipeline implementation to change less of sequence labeling pipeline
